requirements.txt
absl-py==0.13.0
anago==1.0.8
anyio==3.2.1
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
blis==0.7.4
boto3==1.17.106
botocore==1.20.106
bs4==0.0.1
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.2.2
colorama==0.4.4
conllu==4.4
cycler==0.10.0
cymem==2.0.5
# Cython==0.29.23
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
editdistance==0.5.3
entrypoints==0.3
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
ftfy==6.0.3
gast==0.4.0
gensim==3.8.3
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
grpcio==1.34.1
h5py==3.1.0
idna==2.10
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.5
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
overrides==3.1.0
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.0
Penman==0.6.2
pickleshare==0.7.5
Pillow==8.3.1
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycorenlp==0.3.0
pycparser==2.20
pydantic==1.7.4
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-crfsuite == 0.9.7
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
s3transfer==0.4.2
Sastrawi==1.0.1
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.9.0
Send2Trash==1.7.1
seqeval==1.2.2
six==1.15.0
smart-open==5.1.0
sniffio==1.2.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.7
srsly==2.4.1
stanfordnlp==0.2.0
stanza==1.2.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.2
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
thinc==8.0.7
threadpoolctl==2.1.0
# torch==1.9.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.7.4.3
typing-utils==0.1.0
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
word2number==1.1
wordcloud==1.8.1
wrapt==1.12.1
xgboost==1.2.0
